+++ android/src/main/java/io/appmetrica/analytics/push/flutter/impl/LaunchIntentHolder.java
@@ -1,0 +1,10 @@
+package io.appmetrica.analytics.push.flutter.impl;
+
+import android.content.Intent;
+import androidx.annotation.Nullable;
+
+public class LaunchIntentHolder {
+
+    @Nullable
+    public Intent initialIntent;
+}