+++ ios/AppMetricaPushPlugin/AMPFAppMetricaPushInfoConverter.h
@@ -1,0 +1,8 @@
+
+#import "AMPFPigeon.h"
+
+@interface AMPFAppMetricaPushInfoConverter : NSObject
+
++ (AMPFAppMetricaPushInfoPigeon *)toPigeon:(NSDictionary *)userInfo;
+
+@end