+++ lib/src/appmetrica_push_info.dart
@@ -1,0 +1,16 @@
+import 'appmetrica_push_api_pigeon.dart';
+
+class AppMetricaPushInfo {
+
+  final String? payload;
+
+  AppMetricaPushInfo({
+    this.payload
+  });
+
+  factory AppMetricaPushInfo.fromPigeon(AppMetricaPushInfoPigeon pigeon) {
+    return AppMetricaPushInfo(
+      payload: pigeon.payload,
+    );
+  }
+}